+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.1.3] - 2026-07-24
+
 ### Added
 
 - CLI: new repeatable `--ignore <family/name>[:<reason>]` option (also